u9efn3qx9

+++ BlazorApp6/Models/AiEvaluation.cs
@@ -4,7 +4,7 @@
     {
         public Guid Id { get; set; } = Guid.NewGuid();
         public Guid SwapId { get; set; }
-        public double ExplanationClarity { get; set; }
-        public double UnderstandingLevel { get; set; }
+        public int ExplanationClarity { get; set; }
+        public int UnderstandingLevel { get; set; }
     }
 }
